+++ src/templates/webpack/dev.ts
@@ -7,7 +7,7 @@ export default function (config: {
   ts: boolean;
 }) {
   const { project_type, name, style, ts } = config;
-  const isReactSPAProject = project_type === 'spa_react';
+  const isReactSPAProject = project_type === 'spa-react';
 
   return `'use strict';